+++ backend-kt/core/src/main/kotlin/hjj/common/tree/TreeSearchResult.kt
@@ -1,0 +1,7 @@
+package hjj.common.tree
+
+data class TreeSearchResult<T>(
+    val node: T,
+    val ancestors: List<T>,
+    val descendants: List<T>
+)